Rename com.example to com.martgant
