+++ data/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.data"
+        applicationId "com.martgant.data"
         minSdk 21
         targetSdk 31
         versionCode 1
