+++ data/src/androidTest/java/com/martgant/data/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.tolkienquizapp
+package com.martgant.data
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.tolkienquizapp", appContext.packageName)
+        assertEquals("com.martgant.data", appContext.packageName)
     }
 }
