+++ data/src/test/java/com/martgant/data/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.domain
+package com.martgant.data
 
 import org.junit.Test
