+++ domain/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.domain"
+        applicationId "com.martgant.domain"
         minSdk 21
         targetSdk 31
         versionCode 1
