+++ presentation/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.tolkienquizapp"
+        applicationId "com.martgant.tolkienquizapp"
         minSdk 21
         targetSdk 31
         versionCode 1
