+++ presentation/src/main/java/com/martgant/tolkienquizapp/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.tolkienquizapp
+package com.martgant.tolkienquizapp
 
 import android.os.Bundle
 import com.google.android.material.snackbar.Snackbar
@@ -9,7 +9,7 @@ import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import android.view.Menu
 import android.view.MenuItem
-import com.example.tolkienquizapp.databinding.ActivityMainBinding
+import com.martgant.tolkienquizapp.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
